requirements.txt
streamlit
transformers
torch
pandas
numpy
plotly
wordcloud
matplotlib
requests
python-dotenv
